Refactor and modularize the codebase
Refactoring for better design and readability
