+++ propro/constants.py
@@ -1,0 +1,3 @@
+# Constants module
+# TODO: should probably be exposed via the GUI?
+BOX_SPACING = 2
