+++ propro/utils.py
@@ -1,0 +1,3 @@
+def flatten(xss):
+  '''Flatten nested list'''
+  return [x for xs in xss for x in xs]
